Split primary intent runtime helpers
