+++ crates/oasis7/src/viewer/runtime_live/gameplay_snapshot.rs
@@ -12,6 +12,8 @@ use crate::viewer::ACTION_CLAIM_FIRST_AGENT;
 
 #[path = "gameplay_snapshot_fallback.rs"]
 mod fallback;
+#[path = "gameplay_snapshot_intent.rs"]
+mod intent;
 
 use super::branch_commitment::{branch_recommendations, effective_branch_stage_status};
 pub(super) use super::gameplay_snapshot_feedback::player_gameplay_feedback_from_control_ack;
@@ -26,6 +28,7 @@ use fallback::{
     fallback_tradeoff_decision_for_gameplay, player_gameplay_fallback_action,
     player_gameplay_fallback_tradeoff_preview,
 };
+use intent::{player_gameplay_intent_scope, player_gameplay_intent_summary};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(super) struct PlayerGameplayCausalitySignal {
@@ -90,32 +93,6 @@ pub(super) fn player_gameplay_causality_from_runtime_events(
     override_fallback.map(|fallback| PlayerGameplayCausalitySignal {
         kind: PlayerGameplayCausalityKind::AgentOverride,
         detail: override_detail.unwrap_or(fallback),
-    })
-}
-
-fn player_gameplay_intent_scope(action: &str) -> Option<&'static str> {
-    if action.starts_with("gameplay_action:") {
-        Some("gameplay_action")
-    } else if action.starts_with("prompt_control.") {
-        Some("prompt_control")
-    } else if action == "agent_chat" {
-        Some("agent_chat")
-    } else if matches!(action, "play" | "pause" | "step" | "seek") {
-        Some("world_control")
-    } else if action == "chain_sync" {
-        Some("world_sync")
-    } else {
-        None
-    }
-}
-
-fn player_gameplay_intent_summary(feedback: &PlayerGameplayRecentFeedback) -> Option<String> {
-    feedback.intent_summary.clone().or_else(|| {
-        if feedback.action.is_empty() {
-            None
-        } else {
-            Some(feedback.action.replace('_', " "))
-        }
     })
 }
